requirements.txt
asgiref==3.2.7
Django==3.0.4
django-appconf==1.0.4
django-ckeditor==5.9.0
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.1
